+++ src/apps/plutus/arbitrum/plutus.farm-pls-jones.contract-position-fetcher.ts
@@ -2,7 +2,11 @@ import { Inject } from '@nestjs/common';
 
 import { APP_TOOLKIT, IAppToolkit } from '~app-toolkit/app-toolkit.interface';
 import { PositionTemplate } from '~app-toolkit/decorators/position-template.decorator';
-import { GetDataPropsParams, GetTokenBalancesParams } from '~position/template/contract-position.template.types';
+import {
+  GetDataPropsParams,
+  GetDisplayPropsParams,
+  GetTokenBalancesParams,
+} from '~position/template/contract-position.template.types';
 import {
   SingleStakingFarmDataProps,
   SingleStakingFarmDefinition,
@@ -11,9 +15,17 @@ import {
 
 import { PlutusContractFactory, PlutusFarmPlsJones } from '../contracts';
 
+export type PlutusFarmDefinition = SingleStakingFarmDefinition & {
+  label: string;
+};
+
 @PositionTemplate()
-export class ArbitrumPlutusFarmPlsJonesContractPositionFetcher extends SingleStakingFarmTemplateContractPositionFetcher<PlutusFarmPlsJones> {
-  groupLabel = 'plsJONES Farm';
+export class ArbitrumPlutusFarmPlsJonesContractPositionFetcher extends SingleStakingFarmTemplateContractPositionFetcher<
+  PlutusFarmPlsJones,
+  SingleStakingFarmDataProps,
+  PlutusFarmDefinition
+> {
+  groupLabel = 'Farms';
 
   constructor(
     @Inject(APP_TOOLKIT) protected readonly appToolkit: IAppToolkit,
@@ -26,11 +38,12 @@ export class ArbitrumPlutusFarmPlsJonesContractPositionFetcher extends SingleSta
     return this.contractFactory.plutusFarmPlsJones({ address, network: this.network });
   }
 
-  async getFarmDefinitions(): Promise<SingleStakingFarmDefinition[]> {
+  async getFarmDefinitions(): Promise<PlutusFarmDefinition[]> {
     return [
       {
         address: '0x23b87748b615096d1a0f48870daee203a720723d',
         stakedTokenAddress: '0xe7f6c3c1f0018e4c08acc52965e5cbff99e34a44',
+        label: 'plsJONES',
         rewardTokenAddresses: [
           '0x51318b7d00db7acc4026c88c3952b66278b6a67f', // PLS
           '0xf236ea74b515ef96a9898f5a4ed4aa591f253ce1', // plsDPX
@@ -50,6 +63,25 @@ export class ArbitrumPlutusFarmPlsJonesContractPositionFetcher extends SingleSta
 
     const emissions = await rewardsDistroContract.getEmissions();
     return [emissions.pls_, emissions.plsDpx_, emissions.plsJones_, emissions.jones_];
+  }
+
+  async getIsActive({
+    contract,
+  }: GetDataPropsParams<PlutusFarmPlsJones, SingleStakingFarmDataProps, PlutusFarmDefinition>): Promise<boolean> {
+    const rewardsDistro = await contract.rewardsDistro();
+    const rewardsDistroContract = this.contractFactory.plutusRewardsDistroPlsJones({
+      address: rewardsDistro,
+      network: this.network,
+    });
+
+    const emissions = await rewardsDistroContract.getEmissions();
+    return emissions.pls_.gt(0) || emissions.plsDpx_.gt(0) || emissions.plsJones_.gt(0) || emissions.jones_.gt(0);
+  }
+
+  async getLabel({
+    definition,
+  }: GetDisplayPropsParams<PlutusFarmPlsJones, SingleStakingFarmDataProps, PlutusFarmDefinition>) {
+    return definition.label;
   }
 
   async getStakedTokenBalance({ contract, address }: GetTokenBalancesParams<PlutusFarmPlsJones>) {