+++ src/apps/plutus/arbitrum/plutus.farm.contract-position-fetcher.ts
@@ -1,0 +1,71 @@
+import { Inject } from '@nestjs/common';
+import { BigNumberish } from 'ethers';
+
+import { APP_TOOLKIT, IAppToolkit } from '~app-toolkit/app-toolkit.interface';
+import { PositionTemplate } from '~app-toolkit/decorators/position-template.decorator';
+import {
+  GetMasterChefDataPropsParams,
+  GetMasterChefTokenBalancesParams,
+  MasterChefTemplateContractPositionFetcher,
+  RewardRateUnit,
+} from '~position/template/master-chef.template.contract-position-fetcher';
+
+import { PlutusChef, PlutusContractFactory } from '../contracts';
+
+@PositionTemplate()
+export class ArbitrumPlutusFarmContractPositionFetcher extends MasterChefTemplateContractPositionFetcher<PlutusChef> {
+  chefAddress = '0x5593473e318f0314eb2518239c474e183c4cbed5';
+  groupLabel = 'Farms';
+  rewardRateUnit = RewardRateUnit.SECOND;
+
+  constructor(
+    @Inject(APP_TOOLKIT) protected readonly appToolkit: IAppToolkit,
+    @Inject(PlutusContractFactory) protected readonly contractFactory: PlutusContractFactory,
+  ) {
+    super(appToolkit);
+  }
+
+  getContract(address: string): PlutusChef {
+    return this.contractFactory.plutusChef({ address, network: this.network });
+  }
+
+  async getPoolLength(contract: PlutusChef): Promise<BigNumberish> {
+    return contract.poolLength();
+  }
+
+  async getStakedTokenAddress(contract: PlutusChef, poolIndex: number): Promise<string> {
+    return contract.poolInfo(poolIndex).then(v => v.lpToken);
+  }
+
+  async getRewardTokenAddress(contract: PlutusChef): Promise<string> {
+    return contract.PLS();
+  }
+
+  async getTotalAllocPoints({ contract }: GetMasterChefDataPropsParams<PlutusChef>): Promise<BigNumberish> {
+    return contract.totalAllocPoint();
+  }
+
+  async getTotalRewardRate({ contract }: GetMasterChefDataPropsParams<PlutusChef>): Promise<BigNumberish> {
+    return contract.plsPerSecond();
+  }
+
+  async getPoolAllocPoints({ contract, definition }: GetMasterChefDataPropsParams<PlutusChef>): Promise<BigNumberish> {
+    return contract.poolInfo(definition.poolIndex).then(v => v.allocPoint);
+  }
+
+  async getStakedTokenBalance({
+    address,
+    contract,
+    contractPosition,
+  }: GetMasterChefTokenBalancesParams<PlutusChef>): Promise<BigNumberish> {
+    return contract.userInfo(contractPosition.dataProps.poolIndex, address).then(v => v.amount);
+  }
+
+  async getRewardTokenBalance({
+    address,
+    contract,
+    contractPosition,
+  }: GetMasterChefTokenBalancesParams<PlutusChef>): Promise<BigNumberish> {
+    return contract.pendingPls(contractPosition.dataProps.poolIndex, address);
+  }
+}